Resolve conflicts
Former-commit-id: 5574924de0bcbcb8a1d3767686ac4af0910e9583

+++ Intersect Client/Classes/Networking/PacketHandler.cs
@@ -1152,8 +1152,8 @@ namespace Intersect_Client.Classes.Networking
             bf.WriteBytes(packet);
             if (Globals.Me != null)
             {
-                Globals.Me.Experience = bf.ReadLong();
-                Globals.Me.ExperienceToNextLevel = bf.ReadLong();
+                Globals.Me.Experience = bf.ReadInteger();
+                Globals.Me.ExperienceToNextLevel = bf.ReadInteger();
             }
             bf.Dispose();
         }
